Overhaul README.md and the repo structure
In addition to fixing the `README.md` so that it provides a more
helpful description of the project, the pdf of the assignment
has also been included inside `asgmnt_docs` directory for archival
purpose.

In this commit, the `.idea` directory is also included until further
determination on its importance to the overall project.

The last but not least, the .gitignore file now contains the same
content as that of `benford-law` for the sake of standardization.

 Changes to be committed:
	modified:   .gitignore
	new file:   .idea/.name
	new file:   .idea/misc.xml
	new file:   .idea/modules.xml
	new file:   .idea/vcs.xml
	modified:   README.md
	new file:   asgmnt_docs/asgmnt1.pdf
	new file:   benford-law.iml